Moved UI comparison methods from the trait to a new Abstract Page.
