+++ src/main/groovy/com/sysgears/seleniumbundle/pagemodel/ResultsPage.groovy
@@ -1,21 +1,16 @@
 package com.sysgears.seleniumbundle.pagemodel
 
 import com.codeborne.selenide.*
-import com.sysgears.seleniumbundle.core.pagemodel.AbstractPage
+import com.sysgears.seleniumbundle.core.pagemodel.AbstractPageWithUIComparison
 import com.sysgears.seleniumbundle.core.pagemodel.annotations.StaticElement
 import com.sysgears.seleniumbundle.core.uicomparison.ScreenshotHandler
-import groovy.util.logging.Slf4j
 import io.qameta.allure.Step
 import org.testng.Assert
 
 import static com.codeborne.selenide.Selenide.$
 import static com.codeborne.selenide.Selenide.$$
 
-/**
- * Slf4j annotation is required for each page as it will be used during UI comparison.
- */
-@Slf4j
-class ResultsPage extends AbstractPage<ResultsPage> {
+class ResultsPage extends AbstractPageWithUIComparison<ResultsPage> {
 
     @StaticElement
     private SelenideElement toolsButton = $("#hdtb-tls")
